awesome

+++ bootstrap.sh
@@ -8,7 +8,7 @@ for i in "ackrc" "bashrc" "ctags" "gitconfig" "hgrc" "lessfilter" "mailcap" \
     "pythonrc" "pdbrc" "screenrc" "tmux.conf" "vim" "vimrc" "xbindkeysrc" \
     "Xresources" "Xmodmap" "xprofile"
 do
-    rm $HOME/.$i
+    [ -e $HOME/.$i ] && rm $HOME/.$i
     ln -s $HOME/lib/dotfiles/$i $HOME/.$i
 done
 
@@ -16,6 +16,8 @@ rm $HOME/.config/{awesome,zathura} $HOME/bin
 ln -s $HOME/lib/dotfiles/bin/ $HOME/bin
 ln -s $HOME/lib/dotfiles/awesome/ $HOME/.config/awesome
 ln -s $HOME/lib/dotfiles/zathura/ $HOME/.config/zathura
+
+ln -s $HOME/lib/dotfiles/compton.conf $HOME/.config/compton.conf
 
 rm $HOME/.config/ranger/rc.conf
 ln -s $HOME/lib/dotfiles/ranger/rc.conf $HOME/.config/ranger/rc.conf
